verifica o autor do tópico para excluir

+++ src/main/java/hub/forum/api/service/TopicoService.java
@@ -3,17 +3,20 @@ package hub.forum.api.service;
 import hub.forum.api.domain.curso.Curso;
 import hub.forum.api.dto.AtualizacaoTopicoDto;
 import hub.forum.api.dto.TopicoDto;
+import hub.forum.api.infra.SecurityFilter;
 import hub.forum.api.repository.CursoRepository;
 import hub.forum.api.repository.TopicoRepository;
 import hub.forum.api.repository.UsuarioRepository;
 import hub.forum.api.domain.topico.DadosCadastroTopico;
 import hub.forum.api.domain.topico.Topico;
 import hub.forum.api.domain.usuario.Usuario;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,6 +36,7 @@ public class TopicoService {
 
     @Autowired
     private UsuarioRepository usuarioRepository;
+
 
 
 
@@ -80,16 +84,19 @@ public class TopicoService {
    }
 
    @Transactional
-   public void deletarTopico(Long id){
+   public void deletarTopico(Long id, Long idUsuario){
         Optional<Topico> optionalTopico = topicoRepository.findById(id);
         if (optionalTopico.isPresent()){
-            topicoRepository.deleteById(id);
+            Topico topico = optionalTopico.get();
+            if (topico.getUsuario().getId().equals(idUsuario)){
+                topicoRepository.deleteById(id);
+            }else{
+                throw new IllegalArgumentException("Usuário não autorizado a deletar esse tópico");
+            }
         }else{
             throw new IllegalArgumentException("Tópico não encontrado!");
         }
    }
-
-
